backend/requirements-python313.txt
# FastAPI and server - Updated versions with Python 3.13 wheels
fastapi==0.115.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Database - Using asyncpg (pure Python, works on Python 3.13)
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary>=2.9.9  # PostgreSQL driver
asyncpg==0.30.0

# Authentication - Using PyJWT instead of python-jose (no Rust needed)
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
cryptography>=44.0.0  # Pre-built wheels available for Python 3.13

# LLM Integration
httpx==0.27.2  # Compatible with ollama
ollama==0.4.5

# Utilities
python-dotenv==1.0.1
email-validator==2.2.0

# Testing (optional, for development)
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
